Compilation Try
C++ Compilation is a nightmare. It's incredible how impossible it is to arrange your code freely. The Compiler is a real piece of garbage, and you need to think through all your damn files, to make sure he's able to speak the same damn language as you. This is completely fucking annoying.